Split Web.Stripe.Types to sub modules
