+++ stripe-core/src/Web/Stripe/Types/Country.hs
@@ -1,0 +1,10 @@
+{-# LANGUAGE DeriveDataTypeable #-}
+module Web.Stripe.Types.Country where
+
+import           Data.Data (Data, Typeable)
+import           Data.Text (Text)
+
+------------------------------------------------------------------------------
+-- | Country
+newtype Country       =
+  Country Text deriving (Read, Show, Eq, Ord, Data, Typeable)
